Included basic auth for journal apis also

+++ src/main/java/com/akhilesh/journal_app/service/JournalEntryService.java
@@ -2,6 +2,8 @@ package com.akhilesh.journal_app.service;
 
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.stream.Collectors;
+
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -31,7 +33,7 @@ public class JournalEntryService {
       journalEntry.setDate(LocalDateTime.now());
       JournalEntry savedEntry = repo.save(journalEntry);
       user.getJournalEntries().add(savedEntry);
-      userService.saveEntry(user);
+      userService.saveUser(user);
     }
 
     public void saveEntry(JournalEntry journalEntry){
@@ -42,22 +44,35 @@ public class JournalEntryService {
       return repo.findById(id).orElse(null);
     }
 
-    public void deleteJournalEntryById(ObjectId id, String userName){
+    public boolean deleteJournalEntryById(ObjectId id, String userName){
       User user = userService.findByUserName(userName);
-      user.getJournalEntries().removeIf(x->x.getId().equals(id));
-      userService.saveEntry(user);
-      repo.deleteById(id);
+      Boolean removed = user.getJournalEntries().removeIf(x->x.getId().equals(id));
+      if(removed)
+      { 
+        userService.saveUser(user);
+        repo.deleteById(id);
+      }
+      return removed;
     }
 
     public ResponseEntity<?> updateJournalEntryById(ObjectId myId, JournalEntry myEntry, String userName) {
-      JournalEntry j = getJournalEntryById(myId);
-      if(j!=null)
+      User user = userService.findByUserName(userName);
+      List<JournalEntry> collect = user.getJournalEntries().stream().filter(x->x.getId().equals(myId)).collect(Collectors.toList());
+      if(!collect.isEmpty())
       {
-        j.setTitle(myEntry.getTitle()!=null&&!myEntry.getTitle().equals("") ? myEntry.getTitle():j.getTitle());
-        j.setContent(myEntry.getContent()!=null&&!myEntry.getContent().equals("") ? myEntry.getContent():j.getContent());
-        saveEntry(j);
-        return new ResponseEntity<>(j, HttpStatus.OK);
+        JournalEntry j = getJournalEntryById(myId);
+        if(j!=null)
+        {
+          j.setTitle(myEntry.getTitle()!=null&&!myEntry.getTitle().equals("") ? myEntry.getTitle():j.getTitle());
+          j.setContent(myEntry.getContent()!=null&&!myEntry.getContent().equals("") ? myEntry.getContent():j.getContent());
+          saveEntry(j);
+          return new ResponseEntity<>(j, HttpStatus.OK);
+        }
       }
       return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+    }
+
+    public List<JournalEntry> findByUserName(String userName){
+      return null;
     }
 }
